Keep channel tuning fractional, and fix the drum and key-shift layers
RPN fine tuning resolves to 100/8192 of a cent, and the port was throwing
almost all of that away.  ss_channel_set_tuning rounded its argument to whole
cents, and channel_tuning_cents was an int that truncated the sum of the tuning
custom controllers again; render_voice then narrowed what was left into a float
accumulator.  Upstream holds every one of those as a number.  All three are
widened, and the two roundings are gone.

Two neighbouring bugs in the same computation.  A drum channel takes the
channel *system* fine tune and key shift alone -- the port was adding the
channel MIDI layer to it as well.  And the key shift is truncated toward zero,
as upstream's Math.trunc does, not floored, which sent a fractional negative
shift a semitone further down than intended.

rpn_fine_tuning_test 37.50 -> 118.44 dB.  Nothing else moves, the three C tests
pass, and near-identical rises to seventeen of thirty-one.

rpn_tuning_real-time_test is untouched at 6.28 dB and is a different bug: RPN
coarse tuning should set the channel's key shift, which reselects the sample
zone, where the port applies it as a semitone tuning offset.  The test says so
itself -- "should be treated as non-realtime (key shift)".

+++ spessasynth_core/include/spessasynth/synthesizer/synth.h
@@ -490,7 +490,7 @@ typedef struct SS_MIDIChannel {
 	SS_DynamicModulatorSystem dms;
 
 	int8_t channel_octave_tuning[128];
-	int channel_tuning_cents;
+	double channel_tuning_cents;
 	int16_t generator_offsets[SS_GEN_COUNT];
 	bool generator_offsets_enabled;
 	int16_t generator_overrides[SS_GEN_COUNT];
